Scalar::FromArraySlot -> Array::GetScalar
