+++ cpp/src/arrow/scalar.cc
@@ -270,100 +270,6 @@ Result<std::shared_ptr<Scalar>> Scalar::Parse(const std::shared_ptr<DataType>& t
   return ScalarParseImpl{type, s}.Finish();
 }
 
-struct ScalarFromArraySlotImpl {
-  template <typename T>
-  using ScalarType = typename TypeTraits<T>::ScalarType;
-
-  Status Visit(const NullArray& a) {
-    out_ = std::make_shared<NullScalar>();
-    return Status::OK();
-  }
-
-  Status Visit(const BooleanArray& a) { return Finish(a.Value(index_)); }
-
-  template <typename T>
-  Status Visit(const NumericArray<T>& a) {
-    return Finish(a.Value(index_));
-  }
-
-  template <typename T>
-  Status Visit(const BaseBinaryArray<T>& a) {
-    return Finish(a.GetString(index_));
-  }
-
-  Status Visit(const FixedSizeBinaryArray& a) { return Finish(a.GetString(index_)); }
-
-  Status Visit(const DayTimeIntervalArray& a) { return Finish(a.Value(index_)); }
-
-  template <typename T>
-  Status Visit(const BaseListArray<T>& a) {
-    return Finish(a.value_slice(index_));
-  }
-
-  Status Visit(const FixedSizeListArray& a) { return Finish(a.value_slice(index_)); }
-
-  Status Visit(const StructArray& a) {
-    ScalarVector children;
-    for (const auto& child : a.fields()) {
-      children.emplace_back();
-      ARROW_ASSIGN_OR_RAISE(children.back(), Scalar::FromArraySlot(*child, index_));
-    }
-    return Finish(std::move(children));
-  }
-
-  Status Visit(const UnionArray& a) {
-    return Status::NotImplemented("Non-null UnionScalar");
-  }
-
-  Status Visit(const DictionaryArray& a) {
-    ARROW_ASSIGN_OR_RAISE(auto index, Scalar::FromArraySlot(*a.indices(), index_));
-    ARROW_ASSIGN_OR_RAISE(auto index64, index->CastTo(int64()));
-    ARROW_ASSIGN_OR_RAISE(
-        auto value,
-        Scalar::FromArraySlot(*a.dictionary(),
-                              checked_cast<const Int64Scalar&>(*index64).value));
-    return Finish(std::move(value));
-  }
-
-  Status Visit(const ExtensionArray& a) {
-    return Status::NotImplemented("Non-null ExtensionScalar");
-  }
-
-  template <typename Arg>
-  Status Finish(Arg&& arg) {
-    return MakeScalar(array_.type(), std::forward<Arg>(arg)).Value(&out_);
-  }
-
-  Status Finish(std::string arg) {
-    return MakeScalar(array_.type(), Buffer::FromString(std::move(arg))).Value(&out_);
-  }
-
-  Result<std::shared_ptr<Scalar>> Finish() && {
-    if (index_ >= array_.length()) {
-      return Status::IndexError("tried to refer to element ", index_,
-                                " but array is only ", array_.length(), " long");
-    }
-
-    if (array_.IsNull(index_)) {
-      return MakeNullScalar(array_.type());
-    }
-
-    RETURN_NOT_OK(VisitArrayInline(array_, this));
-    return std::move(out_);
-  }
-
-  ScalarFromArraySlotImpl(const Array& array, int64_t index)
-      : array_(array), index_(index) {}
-
-  const Array& array_;
-  int64_t index_;
-  std::shared_ptr<Scalar> out_;
-};
-
-Result<std::shared_ptr<Scalar>> Scalar::FromArraySlot(const Array& array, int64_t index) {
-  return ScalarFromArraySlotImpl{array, index}.Finish();
-}
-
 namespace internal {
 Status CheckBufferLength(const FixedSizeBinaryType* t, const std::shared_ptr<Buffer>* b) {
   return t->byte_width() == (*b)->size()
